GitHub workflows update (#100)
* upgrade github workflows to sbf, and make the checks pass

* second try to make checks pass

* third try to fix checks, add draft sec3 audits

* fix token-haver get_space()

* recommit to run sec3 audit

+++ programs/nft-voter/src/instructions/update_voter_weight_record.rs
@@ -58,7 +58,7 @@ pub fn update_voter_weight_record(
             &mut unique_nft_mints,
         )?;
 
-        voter_weight = voter_weight.checked_add(nft_vote_weight as u64).unwrap();
+        voter_weight = voter_weight.checked_add(nft_vote_weight).unwrap();
     }
 
     let voter_weight_record = &mut ctx.accounts.voter_weight_record;
